Consolidation and cleanup of energy classes into sst_base. Untested
